Worked on registration feature
Feature is working, still need to do some finishing touches. I still need to add error checking to the functionality. Some other are fetaures are 'on' due to testing purposes. There is a clear communication between the api and the mobile application.

Mac is still giving issues when trying to test iOS.